removed all instances of heka.senders from tests and code
