+++ heka/streams/__init__.py
@@ -14,8 +14,9 @@
 # ***** END LICENSE BLOCK *****
 
 
+from heka.streams.dev import DebugCaptureStream  # NOQA
 from heka.streams.dev import FileStream  # NOQA
 from heka.streams.dev import StdOutStream  # NOQA
-from heka.streams.dev import DebugCaptureStream  # NOQA
-from heka.streams.udp import UdpStream  # NOQA
+from heka.streams.logging import StdLibLoggingStream # NOQA
 from heka.streams.tcp import TcpStream  # NOQA
+from heka.streams.udp import UdpStream  # NOQA
