+++ heka/tests/test_filters.py
@@ -11,9 +11,10 @@
 #   Rob Miller (rmiller@mozilla.com)
 #
 # ***** END LICENSE BLOCK *****
+from heka.config import build_sender
 from heka.client import HekaClient
 from heka.client import SEVERITY
-from heka.senders import DebugCaptureSender
+from heka.streams import DebugCaptureStream
 from heka.tests.helpers import decode_message
 from nose.tools import eq_, ok_
 import random
@@ -24,7 +25,8 @@ class TestHekaClientFilters(object):
     timer_name = 'test'
 
     def setUp(self):
-        self.sender = DebugCaptureSender()
+        self.stream = DebugCaptureStream()
+        self.sender = build_sender(self.stream, 'heka.encoders.JSONEncoder')
         self.client = HekaClient(self.sender, self.logger)
 
     def tearDown(self):
@@ -42,9 +44,9 @@ class TestHekaClientFilters(object):
         self.client.exception(payload)
         self.client.critical(payload)
         # only half of the messages should have gone out
-        eq_(len(self.sender.msgs), 3)
+        eq_(len(self.stream.msgs), 3)
         # make sure it's the right half
-        for json_msg in self.sender.msgs:
+        for json_msg in self.stream.msgs:
             msg = self._extract_msg(json_msg)
             ok_(msg.severity <= SEVERITY.ERROR)
 
@@ -59,7 +61,7 @@ class TestHekaClientFilters(object):
             if choice != 'foo':
                 notfoos += 1
             self.client.heka(choice, payload='msg')
-        eq_(len(self.sender.msgs), notfoos)
+        eq_(len(self.stream.msgs), notfoos)
 
     def test_type_whitelist(self):
         from heka.filters import type_whitelist_provider
@@ -72,7 +74,7 @@ class TestHekaClientFilters(object):
             if choice == 'foo':
                 foos += 1
             self.client.heka(choice, payload='msg')
-        eq_(len(self.sender.msgs), foos)
+        eq_(len(self.stream.msgs), foos)
 
     def test_type_severity_max(self):
         from heka.filters import type_severity_max_provider
@@ -85,8 +87,8 @@ class TestHekaClientFilters(object):
         for msgtype in ['foo', 'bar']:
             for sev in range(8):
                 self.client.heka(msgtype, severity=sev, payload='msg')
-        eq_(len(self.sender.msgs), 10)
-        msgs = [self._extract_msg(msg) for msg in self.sender.msgs]
+        eq_(len(self.stream.msgs), 10)
+        msgs = [self._extract_msg(msg) for msg in self.stream.msgs]
         foos = [msg for msg in msgs if msg.type == 'foo']
         eq_(len(foos), 4)
         bars = [msg for msg in msgs if msg.type == 'bar']
